Enable all compiler lints, solve low-hanging fruits

+++ microsim-core/src/main/java/microsim/alignment/multiple/LogitScalingWeightedAlignment.java
@@ -182,7 +182,7 @@ public class LogitScalingWeightedAlignment<T extends Weight> extends AbstractMul
 		
 		if( (error >= allowedError) && enableWarnings) {
 			System.out.println("WARNING: The LogitScalingWeightedAlignment.align() method terminated with an error of " 
-					+ (error/(double)total) + ", which has a greater magnitude than the precision bounds of +/-" + precision + ".  The size of "
+					+ (error/total) + ", which has a greater magnitude than the precision bounds of +/-" + precision + ".  The size of "
 					+ " the filtered agent collection is " + n + " and the number of iterations was  " + count + ".  Check the results of the Logit Scaling Weighted alignment to ensure that "
 					+ "alignment is good enough for the purpose in question, or consider increasing the maximum number of iterations "
 					+ "or the precision!");
